1. added in save/restore functions information about last used method of optimization 2. Fixed bug with relative step in Simplex Norm. 3. Added GP optimization from sklearn package. Valid from version 0.18 of sklearn 4. added tooltips for qtable with names of the devices. 5. added dev mode. 6. fixed bug with names of the algorithms. bug appeared only for scipy version > 0.15

+++ cpbd/high_order.py
@@ -810,7 +810,7 @@ def rk_track_in_field(y0, s_stop, N, energy, mag_field, s_start=0.):
     cmm = speed_of_light
     massElectron = m_e_eV# 0.510998910e+6 #// rest mass of electron
 
-    u = np.zeros((N*9, len(y0)/6))
+    u = np.zeros((N*9, int(len(y0)/6)))
     px = y0[1::6]
     py = y0[3::6]
     dz = h
